Refactor GradePopup to use helper function and CourseInfoButton
Also add analytics for clicking 'Grades'
