+++ src/components/SectionTable/SectionTable.js
@@ -14,13 +14,13 @@ import {
 import { withStyles } from '@material-ui/core/styles';
 // import AlmanacGraph from '../EnrollmentGraph/EnrollmentGraph'; uncomment when we get past enrollment data back
 import CourseInfoBar from './CourseInfoBar';
-import GradeInfoBar from './GradeInfoBar';
 import SectionTableBody from './SectionTableBody';
 import CourseInfoButton from './CourseInfoButton';
 import { Help, Assessment, Assignment } from '@material-ui/icons';
 import ShowChartIcon from '@material-ui/icons/ShowChart';
 import PropTypes from 'prop-types';
 import analyticsEnum from '../../analytics';
+import GradesPopup from './GradesPopup';
 
 const styles = {
     flex: {
@@ -97,14 +97,20 @@ const SectionTable = (props) => {
                         redirectLink={courseDetails.prerequisiteLink}
                     />
                 )}
-
-                <GradeInfoBar
+                <CourseInfoButton
+                    analyticsCategory={analyticsCategory}
+                    anlyticsAction={analyticsEnum.classSearch.actions.CLICK_GRADES}
                     text="Grades"
                     icon={<Assessment />}
-                    deptCode={courseDetails.deptCode}
-                    courseNumber={courseDetails.courseNumber}
-                    isMobileScreen={isMobileScreen}
+                    popupContent={
+                        <GradesPopup
+                            deptCode={courseDetails.deptCode}
+                            courseNumber={courseDetails.courseNumber}
+                            isMobileScreen={isMobileScreen}
+                        />
+                    }
                 />
+
                 <CourseInfoButton
                     analyticsCategory={analyticsCategory}
                     anlyticsAction={analyticsEnum.classSearch.actions.CLICK_PAST_ENROLLMENT}
